Better fix for issue #550
